Makes concurrency test into server_test
