+++ tests/server_test.py
@@ -1,4 +1,6 @@
 import unittest
+from multiprocessing.pool import Pool
+
 import requests
 import re
 from itertools import *
@@ -32,6 +34,22 @@ def jsonrpc_post(url, method, **params):
 
 def nothing():
     pass
+
+
+replace = {
+    'claim_id': fake.sha1,
+    'comment': fake.text,
+    'channel_id': fake.sha1,
+    'channel_name': fake_lbryusername,
+    'signature': fake.uuid4,
+    'parent_id': fake.sha256
+}
+
+
+def create_test_comments(values: iter, **default):
+    vars_combo = chain.from_iterable(combinations(values, r) for r in range(1, len(values) + 1))
+    return [{k: replace[k]() if k in comb else v for k, v in default.items()}
+            for comb in vars_combo]
 
 
 class ServerTest(unittest.TestCase):
@@ -216,23 +234,30 @@ class ListCommentsTest(unittest.TestCase):
         self.assertEqual(response['total_pages'], response_one['total_pages'])
 
 
+class ConcurrentWriteTest(unittest.TestCase):
+    @staticmethod
+    def make_comment(num):
+        return {
+            'jsonrpc': '2.0',
+            'id': num,
+            'method': 'create_comment',
+            'params': {
+                'comment': f'Comment #{num}',
+                'claim_id': '6d266af6c25c80fa2ac6cc7662921ad2e90a07e7',
+            }
+        }
 
+    @staticmethod
+    def send_comment_to_server(params):
+        with requests.post(params[0], json=params[1]) as req:
+            return req.json()
 
-
-replace = {
-    'claim_id': fake.sha1,
-    'comment': fake.text,
-    'channel_id': fake.sha1,
-    'channel_name': fake_lbryusername,
-    'signature': fake.uuid4,
-    'parent_id': fake.sha256
-}
-
-
-def create_test_comments(values: iter, **default):
-    vars_combo = chain.from_iterable(combinations(values, r) for r in range(1, len(values) + 1))
-    return [{k: replace[k]() if k in comb else v for k, v in default.items()}
-            for comb in vars_combo]
-
-
-
+    def test01Concurrency(self):
+        urls = [f'http://localhost:{port}/api' for port in range(5921, 5925)]
+        comments = [self.make_comment(i) for i in range(1, 5)]
+        inputs = list(zip(urls, comments))
+        with Pool(4) as pool:
+            results = pool.map(self.send_comment_to_server, inputs)
+        results = list(filter(lambda x: 'comment_id' in x['result'], results))
+        self.assertIsNotNone(results)
+        self.assertEqual(len(results), len(inputs))
